+++ app.py
@@ -16,6 +16,7 @@ sys.path.insert(0, str(Path(__file__).resolve().parent))
 from features.alerts.view import render_alerts_page
 from features.db.view import render_db_page
 from features.eo.view import render_eo_page
+from features.HQ_DESK.view import render_hq_desk_page
 from features.sar.view import render_sar_page
 from features.statistics.view import render_statistics_page
 from home import render_home_page
@@ -32,6 +33,7 @@ pages = [
     st.Page(render_db_page, title="DB", url_path="db"),
     st.Page(render_alerts_page, title="Alerts", url_path="alerts"),
     st.Page(render_statistics_page, title="Statistics", url_path="statistics"),
+    st.Page(render_hq_desk_page, title="HQ Desk", url_path="hq-desk"),
     st.Page(render_blank_1_page, title="Blank 1", url_path="blank-1"),
 ]